fix trapgen and invert, evertything functions correctly if the trapdoor is composed by constant functions and not polynomials

+++ include/random.h
@@ -7,7 +7,7 @@
 
 #include "common.h"
 
-uint16_t P_random();
+signed int P_random();
 signed int gaussian_random(RR_t mean, RR_t std_dev);
 uint16_t uniform_q_random(int q);
